Stop specifying a codec and mode during file opening
My git history doesn't cover why these settings were in use (so it may have
been required for the last site).
Looking at the help for codecs.open() my mode is redundant (since files are
always opened in binary mode) and encoding is optional too.
Removing encoding does mean I'll end up with whatever is in the files so may
not be the ideal fix - I'll have to see if this causes me problems later on.