Update HTML extractor for current site
Identifying the required code is done slightly differently in this website so
the search pattern has had to be updated to the search code I already had lying
around.

As noted in the code, searching for InstanceBeginEditable and using the parent
of that comment would be a better solution and would work without changes on
both sites.